fix(spec): wave-7 adversarial-review fixes — nullable columns, f64 values, honest claims
Round-1 review: FINDINGS — 2 merge-blocking, 2 wording; all addressed:

- MEDIUM: TsValueColumn.unit/aggregation published non-nullable but
  Option<String> without skip_serializing_if SERIALIZES explicit nulls —
  every listing of a minimal column failed strict consumers. Now
  [string, "null"].
- MEDIUM: TsPoint.values lacked format double — the Go SDK generated
  float32 for f64 wire data (silent precision loss; 1e300-class values
  fail to unmarshal). Now format: double (Go: float64 verified).
- MEDIUM (wording): the TD's "matches the Python SDK contract" claim
  overstated — recorded the known divergence: the hand-written
  timeseries.py facade sends ISO-STRING timestamps (axum 422 + silent
  local-storage fallback), sends aggregation fields to /query (serde
  drops them), never calls /aggregate. Pre-existing; the published spec
  is what makes it visible; facade fix is a follow-up.
- LOW (wording): "collection not found → 500" described an error that
  cannot occur — unknown collections yield 200 EMPTY on query/aggregate
  and success:false on delete (no 404 anywhere). Section comment + TD
  reworded.

All four SDKs regenerated; contract gate 15/15; black clean; go+rust
builds green.

+++ clients/go/proximadb/internal/genrest/genrest.gen.go
@@ -2929,7 +2929,7 @@ type TsPoint struct {
 
 	// Timestamp Epoch milliseconds.
 	Timestamp int64               `json:"timestamp"`
-	Values    *map[string]float32 `json:"values,omitempty"`
+	Values    *map[string]float64 `json:"values,omitempty"`
 }
 
 // TsQueryRequest defines model for TsQueryRequest.
@@ -2946,9 +2946,12 @@ type TsQueryResponse struct {
 
 // TsValueColumn defines model for TsValueColumn.
 type TsValueColumn struct {
+	// Aggregation Serialized as an explicit null when unset.
 	Aggregation *string `json:"aggregation,omitempty"`
 	Name        string  `json:"name"`
-	Unit        *string `json:"unit,omitempty"`
+
+	// Unit Serialized as an explicit null when unset.
+	Unit *string `json:"unit,omitempty"`
 }
 
 // TypedFilter A typed filter for search operations
